requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anomalib==2.0.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.12.3
bleach==6.2.0
blessed==1.20.0
bokeh==3.6.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorcet==3.1.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
datashader==0.17.0
debugpy==1.8.7
decorator==5.1.1
docker-pycreds==0.4.0
docstring_parser==0.16
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.9.0
flatbuffers==25.1.24
fonttools==4.51.0
frozenlist==1.7.0
fsspec==2023.4.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
h11==0.14.0
holoviews==1.20.0
httpcore==1.0.7
httpx==0.28.1
humanfriendly==10.0
idna==3.7
imageio==2.37.0
imgviz==1.7.6
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.28.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.9.0
joblib==1.4.2
jsonargparse==4.40.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
kornia==0.8.1
kornia_rs==0.1.9
labelme==5.6.1
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.14.3
linkify-it-py==2.0.3
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multipledispatch==1.0.0
narwhals==1.28.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.61.0
numpy==1.26.3
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
omegaconf==2.3.0
openai==1.72.0
opencv-python==4.11.0.86
osam==0.2.2
packaging==24.0
pandas==2.2.3
panel==1.6.0
param==2.2.0
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.48
propcache==0.3.2
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyct==0.5.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2024.2
pyviz_comms==3.0.4
PyYAML==6.0.1
pyzmq==26.2.0
QtPy==2.4.2
referencing==0.36.2
requests==2.31.0
rich==14.0.0
rich-argparse==1.7.1
rpds-py==0.23.1
rtpt==0.0.4
scikit-image==0.25.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentry-sdk==1.45.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.12
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
toolz==1.0.0
torch==2.2.2
torchaudio==2.2.2
torchinfo==1.8.0
torchmetrics==1.7.3
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
triton==2.2.0
typeshed_client==2.7.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
uc-micro-py==1.0.3
umap-learn==0.5.7
urllib3==2.2.1
wandb==0.16.6
wcwidth==0.2.13
webencodings==0.5.1
xarray==2025.1.1
xyzservices==2025.1.0
yarl==1.20.1
